Initial 2D refactor
